Added Settings, and changed to a single file
